tasks/misc/colorfee/src/requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
pydantic==2.5.3
yarl==1.9.4
python-dotenv~=1.0.1
pip~=23.0.1
wheel~=0.38.4
redis~=5.0.3
setuptools~=67.6.0
aioredis~=2.0.1